fix: update file upload for image
