+++ sera-front/src/components/app/project/SharedRessources/SharedContainer.tsx
@@ -1,8 +1,8 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { BadgeHelp } from "lucide-react";
+import { Loader2 } from "lucide-react";
 import { useState } from "react";
 import { useParams } from "react-router-dom";
-import { Loader2 } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
 import {
